nvim: Hide line numbers
I like the clean look and the extra horizontal space. I've used it like
this for a while and haven't found myself missing the lines. If I need
to go to line 50 I type `:50` (often happens when referncing from
GitHub). If I want to know what line I'm on I use the powerbar.