kitty nvim zsh bash: *Simplify*
* Comment out unused kitty configuration, make it easier to see what
settings I've *changed* from the default.
* Remove base16 from the shell and from neovim.
* Shell prompts don't include git stuff anymore.

I'm trying to get rid of an annoying bug where the termainal "pauses"
until I hit a key. It's confusing when I'm waiting for a command to
finish, and it just sits there until I type a space or something.

Also, things are just snappier now. :)