ubersicht: Make room for the new mic/camera dot
macOS now notifies whenever the mic or camera are being used by
overlaying a yellow dot in the top-right corner of the screen. I don't
like the behavior, but I couldn't find how to disable it. So this makes
room for that dot in my menu bar.

Also, refresh menu bar every second. Didn't seem to have any impact on
my computer.